initial video matrix / crossbar switcher support
Only the actual core code for networked (TCP/IP-controlled) crossbars
from Lightware and Extron. No UI yet, no serially-attached devices yet.

This is basically a port of CompoKit's dvi_matrix_control.py,
but simplified a bit, and with functionality to query the crossbar's
size (number of inputs and outputs) on the initial connection.