completely changed refresh handling
No more refresh tokens; we now have much more direct control:
Each draw call can optionally return a value that specifies
when it wants to be refreshed. These values ripple up the
hierarchy, and the lowest value is used to refresh the screen.

This way, all clocks can precisely specify when they want
to be refreshed, up to the millisecond. This way, we can now
refresh _exactly_ at the desired time when the clock is
supposed to update, and are not rendering any extra frames.