add facility to request a redraw asynchronously
This is required to ensure that when MPD audio fading (which is handled
in a background thread) stops, we can update the visible button state
immediately.

Technically, this is done by using SDL_PushEvent() (which is guaranteed
to be thread-safe) to simulate a dummy event that causes SDL_WaitEvent()
in the main thread to return and subsequently force a screen redraw.

Also, ensure that a fade-out leaves the master volume at its default level.