MPD: attempt to minimize glitches around fade-outs
Don't restore the volume right after pausing, as MPD will still
play parts of the buffer, and this will be audible when fading
out. Instead, restore volume when playback continues. This isn't
watertight either, as MPD will *still* play a few milliseconds
of the old track from its buffer, but this happens even when
not messing with the volume at all, so it seems to be an inherent
limitation of MPD we'll just have to accept.